Renamed module from PyUVS to maven_iuvs.
