+++ maven_iuvs/download.py
@@ -23,8 +23,8 @@ import paramiko
 
 import numpy as np
 
-from PyUVS.miscellaneous import clear_line
-from PyUVS.search import get_latest_files
+from maven_iuvs.miscellaneous import clear_line
+from maven_iuvs.search import get_latest_files
 
 
 def get_user_paths_filename():
@@ -95,7 +95,7 @@ def setup_user_paths():
     user_paths_file = open(user_paths_py, "x")
 
     user_paths_file.write("# This file automatically generated by"
-                          " PyUVS.download.setup_file_paths\n")
+                          " maven_iuvs.download.setup_file_paths\n")
     user_paths_file.write("l1b_dir = \""+l1b_dir+"\"\n")
     user_paths_file.write("spice_dir = \""+spice_dir+"\"\n")
     user_paths_file.write("iuvs_vm_username = \""+vm_username+"\"\n")
@@ -332,7 +332,7 @@ def sync_data(spice=True, l1b=True,
     #  check if user path data exists and set it if not
     setup_user_paths()
     #  load user path data from file
-    from PyUVS.user_paths import l1b_dir, spice_dir, iuvs_vm_username
+    from maven_iuvs.user_paths import l1b_dir, spice_dir, iuvs_vm_username
     if not os.path.exists(spice_dir):
         raise Exception("Cannot find specified SPICE directory."
                         " Is it accessible?")
@@ -506,7 +506,7 @@ def get_euvm_l2b_dir():
         setup_user_paths()
 
     try:
-        from PyUVS.user_paths import euvm_l2b_dir
+        from maven_iuvs.user_paths import euvm_l2b_dir
     except ImportError:
         # need to set euvm_l2b_dir
         euvm_l2b_dir = input("Where should euvm_l2b data be stored?")
@@ -592,7 +592,7 @@ def get_integrated_reports_dir():
         setup_user_paths()
 
     try:
-        from PyUVS.user_paths import integrated_reports_dir
+        from maven_iuvs.user_paths import integrated_reports_dir
     except ImportError:
         # need to set euvm_l2b_dir
         integrated_reports_dir = input("Where should MAVEN Integrated Reports"
