+++ maven_iuvs/geometry.py
@@ -10,8 +10,8 @@ from astropy.io import fits
 from skimage.transform import resize
 import pkg_resources
 
-from PyUVS.instrument import slit_width_deg
-from PyUVS.constants import R_Mars_km
+from maven_iuvs.instrument import slit_width_deg
+from maven_iuvs.constants import R_Mars_km
 
 
 def beta_flip(hdul):
@@ -191,7 +191,7 @@ def highres_swath_geometry(hdul, res=200, twilight='discrete'):
     context_map = np.zeros((hifi_int, hifi_spa, 3))*np.nan
 
     # load Mars surface map and switch longitude domain from [-180,180) to [0, 360)
-    mars_surface_map = plt.imread(os.path.join(pkg_resources.resource_filename('PyUVS', 'ancillary/'),
+    mars_surface_map = plt.imread(os.path.join(pkg_resources.resource_filename('maven_iuvs', 'ancillary/'),
                                                'mars_surface_map.jpg'))
     offset_map = np.zeros_like(mars_surface_map)
     offset_map[:, :1800, :] = mars_surface_map[:, 1800:, :]
