+++ maven_iuvs/time.py
@@ -7,7 +7,7 @@ import pytz
 import spiceypy as spice
 from astropy.io import fits
 
-from PyUVS.search import get_files
+from maven_iuvs.search import get_files
 
 
 def utc_to_sol(utc):
